Automatically backed up by Learn

+++ prime.rb
@@ -4,8 +4,8 @@ def prime?(number)
   last_divider_num = number - 1
   number_array = Array (2..last_divider_num)
   number_array.each do |i|
-    i = i.to_i
-    number = number.to_i
+    i = i.to_f
+    number = number.to_f
     num_divided = number/i
     puts "#{num_divided}"
     puts "#{(num_divided % 1).zero?}"
